Isecl 17327 - Refactor SKC Build and Deployment Scripts

+++ build/skc-tools/sample_apps/build_scripts/build_prerequisites.sh
@@ -1,27 +1,21 @@
 #!/bin//bash
-
-# Check OS and VERSION
-OS=$(cat /etc/os-release | grep ^ID= | cut -d'=' -f2)
-temp="${OS%\"}"
-temp="${temp#\"}"
-OS="$temp"
-VER=$(cat /etc/os-release | grep ^VERSION_ID | tr -d 'VERSION_ID="')
+source ../../config
+if [ $? -ne 0 ]; then
+	echo "unable to read config variables"
+	exit 1
+fi
 
 install_pre_requisites()
 {
 	if [[ "$OS" == "rhel" && "$VER" == "8.1" || "$VER" == "8.2" ]]; then
-		dnf install -qy bc wget tar git gcc-c++ make automake autoconf libtool yum-utils p11-kit-devel cppunit-devel openssl-devel || exit 1
+		$PKGMGR install -qy bc wget tar git gcc-c++ make automake autoconf libtool yum-utils openssl-devel || exit 1
 	elif [[ "$OS" == "ubuntu" && "$VER" == "18.04" ]]; then
-		apt install -y build-essential ocaml ocamlbuild automake autoconf libtool cmake perl libcppunit-dev libssl-dev || exit 1
+		$PKGMGR install -y build-essential ocaml ocamlbuild automake autoconf libtool cmake perl libssl-dev || exit 1
                 wget http://archive.ubuntu.com/ubuntu/pool/main/libt/libtasn1-6/libtasn1-6_4.16.0-2_amd64.deb || exit 1
                 wget http://archive.ubuntu.com/ubuntu/pool/main/libf/libffi/libffi8ubuntu1_3.4~20200819gead65ca871-0ubuntu3_amd64.deb || exit 1
-                wget http://archive.ubuntu.com/ubuntu/pool/main/p/p11-kit/libp11-kit0_0.23.22-1_amd64.deb || exit 1
-                wget http://archive.ubuntu.com/ubuntu/pool/main/p/p11-kit/libp11-kit-dev_0.23.22-1_amd64.deb || exit 1
 
-                apt install -f -y ./libtasn1-6_4.16.0-2_amd64.deb || exit 1
-                apt install -f -y ./libffi8ubuntu1_3.4~20200819gead65ca871-0ubuntu3_amd64.deb || exit 1
-                apt install -f -y ./libp11-kit0_0.23.22-1_amd64.deb || exit 1
-                apt install -f -y ./libp11-kit-dev_0.23.22-1_amd64.deb || exit 1
+                $PKGMGR install -f -y ./libtasn1-6_4.16.0-2_amd64.deb || exit 1
+                $PKGMGR install -f -y ./libffi8ubuntu1_3.4~20200819gead65ca871-0ubuntu3_amd64.deb || exit 1
 
 		rm -rf *.deb
 	else
